Create .env.local
According to course slide 9B 105, this file should be called ".env.local".  It will contain the local and cloud url's for the backend, so all we hae to do is comment out the one we don't want to use and uncomment the other.  Someone the file name changed to ".env".